% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score}
\alias{score}
\title{score}
\usage{
score(balance, numTrans, creditLine)
}
\arguments{
\item{balance}{balance}

\item{numTrans}{number of transactions}

\item{creditLine}{credit line}
}
\description{
runs fraud example benchmark
}

